conformance/bindingTesting/globalObject/cl_webcl_createContext.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following conditions
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCL createContext conformance test.</title>
    <link rel="stylesheet" href="../../../resources/js-test-style.css"/>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
try {
    debug("WebCL :: CreateContext test case.");
    debug("(optional WebCLContextProperties {WebCLPlatform platform, WebCLDevice[]? devices, CLenum deviceType})");

    var wtu = WebCLTestUtils;
    var webCLPlatform = wtu.getPlatform();
    var webCLDevices = wtu.getDevices(webCLPlatform);

    shouldBeType("webcl.createContext();", "WebCLContext");

    // Testing for property: webCLPlatform.
    shouldBeType("webcl.createContext({platform:webCLPlatform});", "WebCLContext");

    // Testing for property: webCLDevices.
    shouldBeType("webcl.createContext({devices:webCLDevices});", "WebCLContext");

    // Testing for different types of devices.
    shouldBeTypeOrException("webcl.createContext({deviceType:webcl.DEVICE_TYPE_CPU});", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("webcl.createContext({deviceType:webcl.DEVICE_TYPE_GPU});", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("webcl.createContext({deviceType:webcl.DEVICE_TYPE_ALL});", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("webcl.createContext({deviceType:webcl.DEVICE_TYPE_DEFAULT});", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("webcl.createContext({deviceType:webcl.DEVICE_TYPE_ACCELERATOR});", "WebCLContext", "DEVICE_NOT_FOUND");

    // Testing sending null as value in the contextProperties.
    shouldBeType("webcl.createContext({platform:null, devices:null, deviceType:null});", "WebCLContext");

    // Testing for Exception.
    shouldThrowExceptionName("webcl.createContext({platform:new Object()});", "INVALID_PLATFORM");
    shouldThrowExceptionName("webcl.createContext({deviceType:invalid_CLenum});", "INVALID_DEVICE");
    shouldThrowExceptionName("webcl.createContext({devices:[1, 2]});", "INVALID_DEVICE");

    // Test for devices from different platforms.
    var webCLPlatforms = webcl.getPlatforms(), webclDevice1, webclDevice2;
    if (webCLPlatforms.length > 1) {
        webCLDevice1 = webCLPlatforms[0].getDevices()[0];
        webCLDevice2 = webCLPlatforms[1].getDevices()[0];
        shouldThrowExceptionName("webcl.createContext({devices:[webCLDevice1, webCLDevice2]});", "INVALID_DEVICE");
    }
    // FIXME :: Need to test DEVICE_NOT_AVAILABLE.
} catch(e) {
    testFailed(e.description);
}
successfullyParsed = true;
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
